requirements.txt
einops >= 0.6.1
PyYAML
torch >= 1.10
datasets >= 2.14.4
numpy >= 1.25.2
transformers >= 4.31.0
tqdm >= 4.66.1
pandas >= 2.0.3
wandb >= 0.15.8
fancy-einsum >= 0.0.3
jaxtyping
rich >= 13.5.2
scikit-learn
accelerate
timeout-decorator
# Sphinx >= 7.2.2
# furo >= 2023.8.19
# fsspec[http]>=2021.05.0